fix: sequence type int -> Long 변경

+++ backend/src/main/java/SANTA/backend/core/mountain/dto/CafeDTO.java
@@ -5,6 +5,6 @@ public record CafeDTO(
         String name,
         String location,
         String imageUrl,
-        int sequence
+        Long sequence
 ) {
 }
